Rename ATLAS incomplete liblapack.a to liblapack_atlas.a
Forwarded: not-needed
Last-Update: 2017-08-08

The liblapack.a created by ATLAS build system is incomplete: it only contains
a small subset of the full LAPACK API (only those optimized by ATLAS).
Hence rename this library to liblapack_atlas.a to avoid a name conflict with
the full liblapack.a.
In particular, note that the "-Ss flapack" option of the configure script does
not work as expected as of ATLAS 3.10.3.
Last-Update: 2017-08-08
Gbp-Pq: Name rename-lapack-atlas.patch