Add a GENERIC machine type
Bug-Debian: https://bugs.debian.org/719355
Forwarded: not-needed
Last-Update: 2017-08-16

In practice this type will have the same effect than the UNKNOWN machine type,
except that it does not try to autodetect the machine CPU.
This is useful when trying to build a generic package on an arch for which
ATLAS does not define a generic machine type. For example, on a armel buildd
with ARMv7 CPU, ATLAS would detect ARMv7 if provided the UNKNOWN type.
Last-Update: 2017-08-16
Gbp-Pq: Name generic.patch